feat: activity-registration 폴더 생성
